requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0.0
scipy>=1.7.0

# Audio processing
librosa>=0.9.2
sounddevice>=0.4.6
soundfile>=0.11.0
pydub

# Machine learning
scikit-learn>=1.0.0
joblib>=1.1.0
torch>=1.13.0
efficientat>=0.0.1

# Lighting control
pyserial>=3.5
PyDMXControl

# Terminal UI
windows-curses>=2.3.0; sys_platform == 'win32'
